Implementation of Sonic (Speech to Speech)
This is a template from aws docs: https://docs.aws.amazon.com/nova/latest/userguide/speech-code-examples.html in the NodeJS Websocket implementation (NodeJS).

I tested this on my browser and it works fine. 

As I don't have a raspberry PI, i cannot write the code and test it. 

This code provides solid foundation for your raspberry pi code, and ultimately should be the same. Take these files and feed them to gpt, and ask it to convert it for the usage of a raspberry pi to save you time.